[ws-proxy]: extract the config from the cmd package
This allows the Config to be exposed for the installer
