+++ components/ws-proxy/cmd/root.go
@@ -5,16 +5,12 @@
 package cmd
 
 import (
-	"encoding/json"
 	"fmt"
 	"os"
 
-	"github.com/spf13/cobra"
-	"golang.org/x/xerrors"
-
 	"github.com/gitpod-io/gitpod/common-go/log"
 	"github.com/gitpod-io/gitpod/common-go/tracing"
-	"github.com/gitpod-io/gitpod/ws-proxy/pkg/proxy"
+	"github.com/spf13/cobra"
 )
 
 var (
@@ -50,50 +46,4 @@ func Execute() {
 func init() {
 	rootCmd.PersistentFlags().BoolVarP(&jsonLog, "json-log", "j", true, "produce JSON log output on verbose level")
 	rootCmd.PersistentFlags().BoolVarP(&verbose, "verbose", "v", false, "Enable verbose JSON logging")
-}
-
-// Config configures this servuce
-type Config struct {
-	Ingress                     proxy.HostBasedIngressConfig      `json:"ingress"`
-	Proxy                       proxy.Config                      `json:"proxy"`
-	WorkspaceInfoProviderConfig proxy.WorkspaceInfoProviderConfig `json:"workspaceInfoProviderConfig"`
-	PProfAddr                   string                            `json:"pprofAddr"`
-	PrometheusAddr              string                            `json:"prometheusAddr"`
-	ReadinessProbeAddr          string                            `json:"readinessProbeAddr"`
-}
-
-// Validate validates the configuration to catch issues during startup and not at runtime
-func (c *Config) Validate() error {
-	if err := c.Ingress.Validate(); err != nil {
-		return err
-	}
-	if err := c.Proxy.Validate(); err != nil {
-		return err
-	}
-	if err := c.WorkspaceInfoProviderConfig.Validate(); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// getConfig loads and validates the configuration
-func getConfig(fn string) (*Config, error) {
-	fc, err := os.ReadFile(fn)
-	if err != nil {
-		return nil, err
-	}
-
-	var cfg Config
-	err = json.Unmarshal(fc, &cfg)
-	if err != nil {
-		return nil, err
-	}
-
-	err = cfg.Validate()
-	if err != nil {
-		return nil, xerrors.Errorf("config validation error: %w", err)
-	}
-
-	return &cfg, nil
 }
